Fix bug preventing compilation on some systems

+++ FRIES_bin/symm_arnoldi_mol.cpp
@@ -23,7 +23,7 @@ struct MyArgs : public argparse::Args {
     uint32_t max_n_dets = kwarg("max_dets", "Maximum number of determinants on a single MPI process");
     std::string ini_path = kwarg("ini_vecs", "Prefix for files containing the vectors with which to initialize the calculation. Files must have names <ini_vecs>dets<xx> and <ini_vecs>vals<xx>, where xx is a 2-digit number ranging from 0 to (num_states - 1), and be text files");
     uint32_t n_states = kwarg("num_states", "Number of states whose energies will be computed");
-    uint16_t n_krylov = kwarg("n_krylov", "Number of multiplications by (1 - \eps H) to include in each iteration").set_default(1000);
+    uint32_t n_krylov = kwarg("n_krylov", "Number of multiplications by (1 - \eps H) to include in each iteration").set_default(1000);
     bool use_npy = flag("use_numpy", "If set, output files will be in numpy (.npy) format. Otherwise, will be in text (.txt) format");
     
     CONSTRUCTOR(MyArgs);
